Complete UDPClientHandler
Add `UDPClientHandler` to handle different encryption types for UDP packets.

* Implement `channelRead0` method to handle different encryption types (basic, advanced, custom) based on the packet's encryption type.
* Add `handleBasicEncryption`, `handleAdvancedEncryption`, and `handleCustomEncryption` methods to process packets with respective encryption types.
* Print packet details (sequence number, acknowledgment number, content) upon receiving a packet.
* Handle exceptions in the `exceptionCaught` method by printing the stack trace and closing the context.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/zerosoft/avolution?shareId=XXXX-XXXX-XXXX-XXXX).
